requirements.txt
beautifulsoup4==4.8.0
dj-database-url==0.5.0
Django~=3.1.5
django-authtools==1.7.0
django-booking==0.7.2
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-braces==1.13.0
django-countries==5.4
django-crispy-forms
django-fontawesome==1.0
django-heroku==0.3.1
django-hosts==3.0
django-hvad==1.8.0
django-libs==2.0.2
django-static-fontawesome==5.0.8.1
django-widget-tweaks==1.4.2
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.0
psycopg2~=2.8.6
psycopg2-binary==2.8.3
python-dateutil~=2.8.1
python-decouple==3.1
pytz~=2020.5
PyYAML==5.1.2
six
soupsieve==1.9.3
sqlparse~=0.4.1
whitenoise==3.3.1
django-multiselectfield
django-environ
coverage
numpy
pandas
asgiref~=3.3.1
setuptools~=49.2.1
openpyxl
pyodbc
django-pyodbc-azure
django-mssql-backend


# python -m pip install bootstrap4
# pip install django-crispy-forms -U
# python manage.py makemigrations margins
# python manage.py sqlmigrate margins 0001
# python manage.py migrate
# python manage.py migrate --database=icps_db
# python manage.py createsuperuser

https://www.youtube.com/watch?v=CkYvpKZyEqI
# python manage.py dbshell   # for testing db connectivity
python manage.py inspectdb --database=icps_db > models.py
python manage.py makemigrations
python manage.py migrate --fake-initial